+++ src/Logger/Log.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Logger;
 
-use DateTime;
+use DateTimeImmutable;
 use ExtendsSoftware\ExaPHP\Logger\Priority\Critical\CriticalPriority;
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
 use Throwable;
@@ -16,14 +16,14 @@ class Log implements LogInterface
      *
      * @param string                 $message
      * @param PriorityInterface|null $priority
-     * @param DateTime|null          $datetime
+     * @param DateTimeImmutable|null $datetime
      * @param mixed[]                $metaData
      * @param Throwable|null         $throwable
      */
     public function __construct(
         private string $message,
         private readonly ?PriorityInterface $priority = null,
-        private readonly ?DateTime $datetime = null,
+        private readonly ?DateTimeImmutable $datetime = null,
         private ?array $metaData = [],
         private ?Throwable $throwable = null,
     ) {
@@ -48,9 +48,9 @@ class Log implements LogInterface
     /**
      * @inheritDoc
      */
-    public function getDateTime(): DateTime
+    public function getDateTime(): DateTimeImmutable
     {
-        return $this->datetime ?? new DateTime();
+        return $this->datetime ?? new DateTimeImmutable();
     }
 
     /**